+++ src/ast/ast.h
@@ -227,6 +227,24 @@ private:
   Sizeof(const Sizeof &other);
 };
 
+class Offsetof : public Expression
+{
+public:
+  DEFINE_ACCEPT
+  DEFINE_LEAFCOPY(Offsetof)
+
+  Offsetof(SizedType record, std::string &field, location loc);
+  Offsetof(Expression *expr, std::string &field, location loc);
+  ~Offsetof();
+
+  SizedType record;
+  Expression *expr;
+  std::string field;
+
+private:
+  Offsetof(const Offsetof &other);
+};
+
 class Map : public Expression {
 public:
   DEFINE_ACCEPT