+++ src/ast/passes/field_analyser.cpp
@@ -139,6 +139,13 @@ void FieldAnalyser::visit(Sizeof &szof)
   resolve_type(szof.argtype);
 }
 
+void FieldAnalyser::visit(Offsetof &ofof)
+{
+  if (ofof.expr)
+    Visit(*ofof.expr);
+  resolve_type(ofof.record);
+}
+
 void FieldAnalyser::visit(AssignMapStatement &assignment)
 {
   Visit(*assignment.map);