+++ src/ast/passes/semantic_analyser.cpp
@@ -1375,6 +1375,33 @@ void SemanticAnalyser::visit(Sizeof &szof)
   resolve_struct_type(szof.argtype, szof.loc);
 }
 
+void SemanticAnalyser::visit(Offsetof &ofof)
+{
+  ofof.type = CreateUInt64();
+  if (ofof.expr)
+  {
+    ofof.expr->accept(*this);
+    ofof.record = ofof.expr->type;
+  }
+  resolve_struct_type(ofof.record, ofof.loc);
+
+  if (!ofof.record.IsRecordTy())
+  {
+    LOG(ERROR, ofof.loc, err_)
+        << "offsetof() 1st argument is not of a record type.";
+  }
+  else if (!bpftrace_.structs.Has(ofof.record.GetName()))
+  {
+    LOG(ERROR, ofof.loc, err_) << "'" << ofof.record << "' does not exist.";
+  }
+  else if (!ofof.record.HasField(ofof.field))
+  {
+    LOG(ERROR, ofof.loc, err_) << "'" << ofof.record << "' "
+                               << "has no field named "
+                               << "'" << ofof.field << "'";
+  }
+}
+
 void SemanticAnalyser::check_stack_call(Call &call, bool kernel)
 {
   call.type = CreateStack(kernel);