+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to
   - [#2559](https://github.com/iovisor/bpftrace/pull/2559)
 - Enable watchpoint support for PowerPC
   - [#2577](https://github.com/iovisor/bpftrace/pull/2577)
+- Add new function, `offsetof`, get the offset of the element in the struct
+  - [#2579](https://github.com/iovisor/bpftrace/pull/2579)
 #### Changed
 - Improve attaching to uprobes with size 0
   - [#2562](https://github.com/iovisor/bpftrace/pull/2562)